feat(theme): add `auto` theme that always uses a users preferred theme
This also changes the default theme from `light` to `auto`, since we now
listen to changes in the users preferred theme in real time. This allows
Rainbow Palette to automatically switch between themes when a user
updates their browsers or phones settings instead of only reading it
once on the first start.