feat(local-ai): use the browsers local AI to generate a palette
This is a proof of concept and uses the experimental `window.ai` API with a model embedded in some Google Chrome browsers. At the time this is only available in Google Chrome Dev builds behind some flags (see https://afficone.com/blog/window-ai-new-chrome-feature-api/ for more information). In other browsers or if the local AI is not enabled this feature will also not show up in Rainbow Palette.